Another attempt.
